chore: refactor routing system using hash router
