+++ package.json
@@ -38,11 +38,12 @@
   },
   "dependencies": {
     "@ricky0123/vad-web": "^0.0.30",
-    "onnxruntime-web": "^1.17.0",
     "dotenv": "^17.4.1",
     "lucide-react": "^1.8.0",
+    "onnxruntime-web": "^1.17.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.14.1",
     "ws": "^8.18.0",
     "zod": "^4.3.6",
     "zustand": "^5.0.0"
