+++ src/renderer/features/active-session/components/SessionPromptComposer.tsx
